[SIIPStitch] Fix signing issue for GOP/GFX/VBT
Unlike other components, GOP driver, PEIM Graphic PEIM and VBT table
are compressed and signed data in BIOS region. In order to merge
a modified component, resigning is required. Therefore, we have to
'restore' the signing code in the stitching process.

The solution is to create a helper script to create signature data
and pre-append it to the section data.

TEST=run unit tests and ensure the output files are identical to
     those created by the original GOP merge tool

Signed-off-by: Jin, Huang <huang.jin@intel.com>