[stitch] Include Gfx PEIM in OBB stitching flow
IntelGraphicsPeim runs post memory stage and part of OBB region.
Therefore, OBB digest should be updated as part of stitching.

Also fixed an issue in generate_section() because cmd is a reference
to GENSEC_SECTION list. When cmd is changed, the value it points to
changed. Fix that by reinitializing cmd every time.

TEST=Run unit test
     and stitch GFX PEIM with EHL IFWI ER23 image and verify booting

Signed-off-by: Jin, Huang <huang.jin@intel.com>