RTC#308208 Use common sub-region options
The bug was originally from an outdated PSE name string due to BIOS change

A common sub-region configuration 'table' is useful for both stitching
and sub-region capsule tool to be in sync. Users can add new sub-region
settings to IP_OPTIONS dict in siip_constants.py to support stitching
or capsule creation unique to their BIOS features.

This is the first step to make the script to support configurable IP
options.

TEST=Run unit test and also verify images created are identical before
     and after this change.

Signed-off-by: Jin, Huang <huang.jin@intel.com>