[Capsule] RTC#311516 Improvement: support calculating variable field size in JSON file
Zero size field allowed to represent auto size field.
If the field size is 0 for file and string fields, the whole file will be read and copied to the buffer.
Test cases added to test_subregion_capsule.py to validate the whole file or string will be read in case of 0 size.

TEST= pytest tests/test_subregion_capsule.py

Signed-off-by: Mohamed ElGohary <mohamedx.elgohary@intel.com>